переделал action осталось убрать коментарии

+++ backend/api/views.py
@@ -1,3 +1,4 @@
+import re
 from django.db.models import Count, Sum
 from django.http import HttpResponse
 from django_filters.rest_framework import DjangoFilterBackend
@@ -92,30 +93,41 @@ class UserViewSet(ModelViewSet):
         permission_classes=[IsAuthenticated]
     )
     def subscribe(self, request, pk):
-        #  запрос
-        if request.method == 'POST':
-            if request.user.id == int(pk):
-                return Response(
-                    {'errors': 'Подписка на себя невозможна.'},
-                    status=status.HTTP_400_BAD_REQUEST
-                )
-            elif (models.Subscription.objects.filter(
-                author=pk, follower=self.request.user).exists()
+        if request.user.is_authenticated:
+            if (
+                models.Subscription.objects.filter(
+                    author=pk, follower=self.request.user).exists()
             ):
+                if request.method == 'DELETE':
+                    models.Subscription.objects.filter(
+                        author=pk, follower=self.request.user).delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
                 return Response(
                     {'errors': 'Подписка уже существует.'},
                     status=status.HTTP_400_BAD_REQUEST
                 )
-            else:
-                author = models.User.objects.get(id=pk)
+            elif request.user.id == int(pk):
+                return Response(
+                    {'errors': 'Подписка на себя невозможна.'},
+                    status=status.HTTP_400_BAD_REQUEST
+                )
+            elif models.User.objects.filter(id=pk).exists():
+                author = models.User.objects.filter(id=pk).annotate(
+                    recipes_count=Count('recipes')).first()
                 models.Subscription.objects.create(
                     author=author, follower=self.request.user)
-                serializer = serializers.UserSerializer(
+                serializer = serializers.SubscriptionSerializer(
                     author, context={'request': request})
-                return Response(serializer.data)
-        models.Subscription.objects.filter(
-            author=pk, follower=self.request.user).delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+                return Response(
+                    serializer.data, status=status.HTTP_201_CREATED)
+            return Response(
+                {'detail': 'Пользователь не существует.'},
+                status=status.HTTP_400_BAD_REQUEST
+                )
+        return Response(
+            {'detail': 'Пользователь не авторизован.'},
+            status=status.HTTP_401_UNAUTHORIZED
+            )
 
 
 class TagViewSet(ReadOnlyModelViewSet):
@@ -136,6 +148,7 @@ class RecipeViewSet(ModelViewSet):
     filter_backends = (DjangoFilterBackend,)
     filterset_class = RecipeFilter
     pagination_class = paginators.CustomPageNumberPaginator
+    permission_classes = [IsAuthenticatedOrReadOnly]
 
     def get_serializer_class(self):
         if self.action == 'list' or self.action == 'retrieve':
@@ -155,77 +168,148 @@ class RecipeViewSet(ModelViewSet):
     @action(
         methods=['post', 'delete'],
         detail=True,
-        url_path='favorite',
-        permission_classes=[IsAuthenticated]
+        
+        # permission_classes=[IsAuthenticated]
     )
+    # перелопатить
     def favorite(self, request, pk):
-        if models.Recipe.objects.filter(id=pk).exists():
-            if request.method == 'POST':
-                serializer = serializers.FavoriteSerializer(
-                    data={'user': request.user.id, 'recipe': pk}
-                )
-                if serializer.is_valid():
-                    serializer.save()
-                    return Response(
-                        serializer.data, status=status.HTTP_201_CREATED)
-                return Response(
-                    serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
             if (
-                not models.Favorite.objects.filter(
-                    user=request.user.id, recipe=pk).exists()
+                models.Favorite.objects.filter(
+                    recipe=pk, user=self.request.user).exists()
             ):
+                if request.method == 'DELETE':
+                    models.Favorite.objects.filter(
+                        recipe=pk, user=self.request.user).delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
                 return Response(
-                    {'errors': 'Рецепт отсутствует в подписке.'},
-                    status=status.HTTP_400_BAD_REQUEST,
+                    {'errors': 'Рецепт уже в избранном.'},
+                    status=status.HTTP_400_BAD_REQUEST
                 )
-            models.Favorite.objects.filter(user=request.user, recipe=pk).delete()
+            # elif request.user.id == int(pk):
+            #     return Response(
+            #         {'errors': 'Подписка на себя невозможна.'},
+            #         status=status.HTTP_400_BAD_REQUEST
+                # )
+            if models.Recipe.objects.filter(id=pk).exists():
+                recipe = models.Recipe.objects.get(id=pk)
+                models.Favorite.objects.create(
+                    recipe=recipe, user=self.request.user)
+                serializer = serializers.FavoriteShoppingCartRecipeSerializer(
+                    recipe, context={'request': request})
+                return Response(
+                    serializer.data, status=status.HTTP_201_CREATED)
             return Response(
-                {'detail': 'Рецепт удалён из подписки.'},
-                status=status.HTTP_204_NO_CONTENT,
-            )
+                {'detail': 'Рецепт не существует.'},
+                status=status.HTTP_400_BAD_REQUEST
+                )
         return Response(
-                {'detail': 'Рецепт не найден.'},
-                status=status.HTTP_400_BAD_REQUEST,
+            {'detail': 'Пользователь не авторизован.'},
+            status=status.HTTP_401_UNAUTHORIZED
             )
+
+        # if models.Recipe.objects.filter(id=pk).exists():
+        #     if request.method == 'POST':
+        #         serializer = serializers.FavoriteSerializer(
+        #             data={'user': request.user.id, 'recipe': pk}
+        #         )
+        #         if serializer.is_valid():
+        #             serializer.save()
+        #             return Response(
+        #                 serializer.data, status=status.HTTP_201_CREATED)
+        #         return Response(
+        #             serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        #     if (
+        #         not models.Favorite.objects.filter(
+        #             user=request.user.id, recipe=pk).exists()
+        #     ):
+        #         return Response(
+        #             {'errors': 'Рецепт отсутствует в подписке.'},
+        #             status=status.HTTP_400_BAD_REQUEST,
+        #         )
+        #     models.Favorite.objects.filter(
+        #         user=request.user, recipe=pk).delete()
+        #     return Response(
+        #         {'detail': 'Рецепт удалён из подписки.'},
+        #         status=status.HTTP_204_NO_CONTENT,
+        #     )
+        # return Response(
+        #         {'detail': 'Рецепт не найден.'},
+        #         status=status.HTTP_400_BAD_REQUEST,
+        #     )
 
     @action(
         methods=['post', 'delete'],
         detail=True,
-        url_path='shopping_cart',
         permission_classes=[IsAuthenticated]
     )
     def shopping_cart(self, request, pk):
-        if check_id(request, pk):
-            id_recipe = request.data.get('id')
-            user = request.user
-            if request.method == 'POST':
-                serializer = serializers.ShoppingCartSerializer(
-                    data={'user': user.id, 'recipe': id_recipe}
-                )
-                if serializer.is_valid():
-                    serializer.save()
-                    return Response(
-                        serializer.data, status=status.HTTP_201_CREATED)
-                return Response(
-                    serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        if request.user.is_authenticated:
             if (
-                not models.ShoppingCart.objects.filter(
-                    user=user.id, recipe=id_recipe).exists()
+                models.ShoppingCart.objects.filter(
+                    recipe=pk, user=self.request.user).exists()
             ):
+                if request.method == 'DELETE':
+                    models.ShoppingCart.objects.filter(
+                        recipe=pk, user=self.request.user).delete()
+                    return Response(status=status.HTTP_204_NO_CONTENT)
                 return Response(
-                    {'errors': 'Рецепт отсутствует в списке покупок.'},
-                    status=status.HTTP_400_BAD_REQUEST,
+                    {'errors': 'Рецепт уже в покупках.'},
+                    status=status.HTTP_400_BAD_REQUEST
                 )
-            models.ShoppingCart.objects.filter(
-                user=user.id, recipe=id_recipe).delete()
+            # elif request.user.id == int(pk):
+            #     return Response(
+            #         {'errors': 'Подписка на себя невозможна.'},
+            #         status=status.HTTP_400_BAD_REQUEST
+                # )
+            if models.Recipe.objects.filter(id=pk).exists():
+                recipe = models.Recipe.objects.get(id=pk)
+                models.ShoppingCart.objects.create(
+                    recipe=recipe, user=self.request.user)
+                serializer = serializers.FavoriteShoppingCartRecipeSerializer(
+                    recipe, context={'request': request})
+                return Response(
+                    serializer.data, status=status.HTTP_201_CREATED)
             return Response(
-                {'detail': 'Рецепт удалён из списка покупок.'},
-                status=status.HTTP_204_NO_CONTENT,
-            )
+                {'detail': 'Рецепт не существует.'},
+                status=status.HTTP_400_BAD_REQUEST
+                )
         return Response(
-                {'detail': 'Неверный id'},
-                status=status.HTTP_404_NOT_FOUND
-        )
+            {'detail': 'Пользователь не авторизован.'},
+            status=status.HTTP_401_UNAUTHORIZED
+            )
+
+        # if check_id(request, pk):
+        #     id_recipe = request.data.get('id')
+        #     user = request.user
+        #     if request.method == 'POST':
+        #         serializer = serializers.ShoppingCartSerializer(
+        #             data={'user': user.id, 'recipe': id_recipe}
+        #         )
+        #         if serializer.is_valid():
+        #             serializer.save()
+        #             return Response(
+        #                 serializer.data, status=status.HTTP_201_CREATED)
+        #         return Response(
+        #             serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        #     if (
+        #         not models.ShoppingCart.objects.filter(
+        #             user=user.id, recipe=id_recipe).exists()
+        #     ):
+        #         return Response(
+        #             {'errors': 'Рецепт отсутствует в списке покупок.'},
+        #             status=status.HTTP_400_BAD_REQUEST,
+        #         )
+        #     models.ShoppingCart.objects.filter(
+        #         user=user.id, recipe=id_recipe).delete()
+        #     return Response(
+        #         {'detail': 'Рецепт удалён из списка покупок.'},
+        #         status=status.HTTP_204_NO_CONTENT,
+        #     )
+        # return Response(
+        #         {'detail': 'Неверный id'},
+        #         status=status.HTTP_404_NOT_FOUND
+        # )
 
     @action(detail=False)
     def download_shopping_cart(self, request):
@@ -238,7 +322,7 @@ class RecipeViewSet(ModelViewSet):
             sum_amount=Sum('recipe_ingredients__amount'))
         serializer = serializers.ShoppingCartDownloadSerializer(
             ingredients, many=True)
-        content = '***СПИСОК ПОКУПОК.***\n\n'
+        content = '***СПИСОК ПОКУПОК***\n\n'
         for val in serializer.data:
             content += '* {}: {} {}\n'.format(
                 val['name'], val['amount'], val['measurement_unit'])
